requirements.txt
torch
torchvision
numpy<2
Pillow
huggingface_hub
einops
safetensors
